+++ chainer/links/normalization/batch_normalization.py
@@ -104,12 +104,12 @@ class BatchNormalization(link.Link):
             gamma = self.gamma
         else:
             gamma = variable.Variable(self.xp.ones(
-                self.avg_mean.shape, dtype=x.data.dtype), volatile='auto')
+                self.avg_mean.shape, dtype=x.dtype), volatile='auto')
         if hasattr(self, 'beta'):
             beta = self.beta
         else:
             beta = variable.Variable(self.xp.zeros(
-                self.avg_mean.shape, dtype=x.data.dtype), volatile='auto')
+                self.avg_mean.shape, dtype=x.dtype), volatile='auto')
 
         if use_batch_mean:
             if finetune:
